intern/Nithyasri.md/fastapi_odoo/requirements.txt
fastapi
uvicorn
requests
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
